basket

+++ phytostore/details.php
@@ -1,16 +1,21 @@
 <?php
-require_once "./includes/header.php";
 require_once "./src/Phyto.php";
 require_once "./src/Effet.php";
 require_once "./src/PhytoModel.php";
 require_once "./src/BasketModel.php";
-
+require_once "./includes/header.php";
 $phytoModel = new PhytoModel();
 $basketModel = new BasketModel();
 $plant = $phytoModel->getOnePlant();
-$basketModel->addToBasket($plant->getId());
-$phytoModel->deletePlant();
 $user = getLoggedUser();
+
+if (isset($_POST['delete'])) {
+    $phytoModel->deletePlant();
+} else if (isset($_POST['add'])) {
+    $basketModel->addToBasket($plant->getId());
+}
+
+
 
 
 
@@ -26,24 +31,23 @@ foreach ($effets as $effet) {
 ?>
 
 
-
-
-<section class="rounded overflow-hidden bg-white shadow-md mb-20">
-
+<section class="rounded overflow-hidden bg-white shadow-md">
     <header class="py-4 bg-green-800 text-white">
         <h1 class="text-6xl text-center font-light uppercase"><?= $plant->getPlante() ?></h1>
     </header>
+
     <div class="flex p-10 gap-6">
         <div class="w-1/3">
             <img class="max-w-full" src="<?= $plant->getImage() ?>" alt="">
         </div>
 
         <div class="w-2/3">
-            <h2 class="text-4xl font-light text-orange-500"><?= $effet->getName() ?></h2>
+            <h2 class="text-4xl font-light text-orange-500"><?= $effetname ?></h2>
+            <br />
             <h3 class="text-4xl text-green-500"><?= $plant->getPrice() ?> €</h3>
 
             <div class="mt-10">
-                <p class="text-2xl text-green-800 font-bold">Description</p>
+                <p class="font-bold text-2xl text-green-700">Description</p>
                 <hr class="my-6" />
                 <p><?= $plant->getDescription() ?></p>
             </div>
@@ -54,7 +58,7 @@ foreach ($effets as $effet) {
                         <i class="fa-solid fa-heart mr-2"></i> J'aime
                     </a>
                     <form method="post">
-                        <button class="text-orange-400 hover:text-orange-600">
+                        <button class="text-orange-400 hover:text-orange-600" name="add">
                             <i class="fa-solid fa-cart-arrow-down mr-2"></i> Ajouter au panier
                         </button>
                     </form>
@@ -66,12 +70,18 @@ foreach ($effets as $effet) {
                             <i class="fa-solid fa-pen-to-square mr-2"></i>Éditer
                         </a>
                         <form method="post" onsubmit="return confirm('Supprimer cette plante ?')">
-                            <button class="py-2 px-4 rounded bg-pink-400 hover:bg-pink-500 text-white">
+                            <button class="py-2 px-4 rounded bg-pink-400 hover:bg-pink-500 text-white" name="delete">
                                 <i class="fa-solid fa-ban mr-2"></i>Supprimer
                             </button>
                         </form>
                     </div>
                 <?php endif ?>
+            <?php else : ?>
+                <div class="mt-10">
+                    <a class="text-orange-400 hover:text-orange-600" href="register.php">
+                        <i class="fa-solid fa-cart-arrow-down mr-2"></i> Ajouter au panier
+                    </a>
+                </div>
             <?php endif ?>
 
             <div class="mt-10 text-right">
@@ -83,8 +93,7 @@ foreach ($effets as $effet) {
     </div>
 
 </section>
-<br />
-
 <?php
-require_once "./components/caroussel_footer.php";
-require_once "./includes/footer.php"; ?>
+require_once "./includes/footer.php";
+
+?>
